requirements.txt
asgiref==3.8.1
autopep8==2.3.0
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
pycodestyle==2.12.0
pymongo==3.12.3
pytz==2024.1
requests==2.32.3
setuptools==70.1.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
